+++ src/components/RegCentersMapTable/tableConfig.js
@@ -1,6 +1,6 @@
 /**@jsx jsx*/
 import {jsx} from "@emotion/core";
-import {TABLE_DATA_URL, FILTER_DATA_URL} from "../../constants/urls";
+import {TABLE_DATA_URL, FILTER_DATA_URL} from "./constants/urls";
 import {DropdownEditor} from "@rsb/simple-table"
 
 const RegCenterEditor = props => <DropdownEditor minWidthOfList={200} {...props}/>
@@ -72,7 +72,7 @@ export const config = {
         },
         {
             accessor: 'office',
-            title: 'Оффис',
+            title: 'Офис',
             minWidth: 100,
             maxWidth: 300,
             sortable: true,