+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 1.0.1
 * first version of react voice app
   * built top menu and reg centers mapping table with editable reg center column
-  
+## 1.0.2
+* added reg centers mapping table with editable rec centers column
+* added a menu item Tools and page for searching unregistered phones in offices